engine:: MGRendererLevelEditor: implement MGScene; MGScene: implement mesh with simple trigger using MGMatrixTriggerMesh
