+++ engine/src/main/java/good/damn/engine/opengl/renderer/MGRendererLevelEditor.kt
@@ -7,7 +7,6 @@ import android.opengl.GLES30.*
 import android.util.Log
 import android.view.MotionEvent
 import good.damn.engine.MGEngine
-import good.damn.engine.MGEngine.Companion
 import good.damn.engine.interfaces.MGIRequestUserContent
 import good.damn.engine.opengl.MGArrayVertex
 import good.damn.engine.opengl.drawers.MGDrawerLightDirectional
@@ -38,6 +37,7 @@ import good.damn.engine.opengl.matrices.MGMatrixTransformationInvert
 import good.damn.engine.opengl.matrices.MGMatrixTransformationNormal
 import good.damn.engine.opengl.matrices.MGMatrixTranslate
 import good.damn.engine.opengl.models.MGMDrawMode
+import good.damn.engine.opengl.scene.MGScene
 import good.damn.engine.opengl.shaders.MGShaderDefault
 import good.damn.engine.opengl.shaders.MGShaderSkySphere
 import good.damn.engine.opengl.shaders.MGShaderSingleMode
@@ -46,6 +46,7 @@ import good.damn.engine.opengl.textures.MGTexture
 import good.damn.engine.opengl.thread.MGHandlerGl
 import good.damn.engine.opengl.triggers.MGDrawerTriggerStateable
 import good.damn.engine.opengl.triggers.MGManagerTriggerState
+import good.damn.engine.opengl.triggers.MGMatrixTriggerMesh
 import good.damn.engine.opengl.triggers.methods.MGTriggerMethodBox
 import good.damn.engine.opengl.triggers.MGTriggerSimple
 import good.damn.engine.touch.MGIListenerScale
@@ -58,14 +59,12 @@ import good.damn.engine.utils.MGUtilsBuffer
 import good.damn.engine.utils.MGUtilsVertIndices
 import java.io.File
 import java.util.concurrent.ConcurrentLinkedQueue
-import kotlin.math.abs
 import kotlin.math.cos
 import kotlin.math.sin
 
 class MGRendererLevelEditor(
     requesterUserContent: MGIRequestUserContent
-): GLSurfaceView.Renderer,
-MGIListenerOnIntersectPosition {
+): GLSurfaceView.Renderer {
 
     companion object {
         private const val TAG = "MGRendererLevelEditor"
@@ -77,228 +76,17 @@ MGIListenerOnIntersectPosition {
     private val mShaderTexCoords = MGShaderSingleMode()
     private val mShaderWireframe = MGShaderSingleMode()
 
-    private val modelMatrixSky = MGMatrixScale().apply {
-        setScale(
-            200000f,
-            200000f,
-            200000f
-        )
-        invalidateScale()
-        invalidatePosition()
-    }
-    private val modelMatrixCamera = MGMatrixTranslate()
-    private val modelMatrixLandscape = MGMatrixTransformationNormal(
-        MGMatrixTranslate(),
-        mShaderDefault
-    )
-
-    private val modelMatrixMeshTest = MGMatrixTransformationNormal(
-        MGMatrixScale(),
-        mShaderDefault
-    )
-
-    private val mVerticesBatchObject = MGArrayVertex()
-
-    private val mVerticesSky = MGArrayVertex()
-    private val mVerticesLandscape = MGArrayVertex()
-
-    private val mGeneratorLandscape = MGGeneratorLandscape(
-        mVerticesLandscape
-    )
-
-    private val materialInteract = MGMaterial(
-        mShaderDefault.material
-    )
-
-    private val materialLandscape = MGMaterial(
-        mShaderDefault.material
-    )
-
-    private val mTextureSky = MGTexture(
-        mShaderSky
-    )
-
-    private val mTextureInteract = MGTexture(
-        mShaderDefault
-    )
-
-    private val mTextureLandscape = MGTexture(
-        mShaderDefault
-    )
-
-    private val mDrawerSwitchBatch = MGDrawerModeSwitch(
-        mVerticesBatchObject,
-        MGDrawerMeshOpaque(
-            mVerticesBatchObject,
-            mTextureInteract,
-            materialInteract
-        )
-    )
-
-    private val meshLandscape = MGMesh(
-        MGDrawerModeSwitch(
-            mVerticesLandscape,
-            MGDrawerMeshOpaque(
-                mVerticesLandscape,
-                mTextureLandscape,
-                materialLandscape
-            ),
-            GL_CW
-        ),
-        mShaderDefault,
-        modelMatrixLandscape.model,
-        modelMatrixLandscape.normal
-    )
-
-    private val meshSky = MGMesh(
-        MGDrawerModeSwitch(
-            mVerticesSky,
-            MGDrawerSkyOpaque(
-                mVerticesSky,
-                mTextureSky
-            ),
-            GL_CCW
-        ),
-        mShaderSky,
-        modelMatrixSky,
-        normals = null
-    )
-
-    private val meshTest = MGMesh(
-        MGDrawerModeSwitch(
-            mVerticesBatchObject,
-            MGDrawerMeshOpaque(
-                mVerticesBatchObject,
-                mTextureLandscape,
-                materialLandscape
-            ),
-            GL_CW
-        ),
-        mShaderDefault,
-        modelMatrixMeshTest.model,
-        modelMatrixMeshTest.normal
-    )
-
-    private val mCameraFree = MGCameraFree(
-        modelMatrixCamera
-    )
-
-    private val mCallbackOnDeltaInteract = MGCallbackOnDeltaInteract()
-    private val mCallbackOnCameraMove = MGCallbackOnCameraMovement(
-        mCameraFree
-    ).apply {
-        setListenerIntersection(
-            this@MGRendererLevelEditor
-        )
-    }
-
-    private val mHandler = MGHandlerGl()
-
-    private val meshes = ConcurrentLinkedQueue<
-        MGDrawerMeshSwitch
-    >().apply {
-        add(meshLandscape)
-        add(meshTest)
-    }
-
-    private val mTriggers = ConcurrentLinkedQueue<
-        MGDrawerTriggerStateable
-    >()
-
     private var mWidth = 0
     private var mHeight = 0
 
-    private val mDrawerLightDirectional = MGDrawerLightDirectional(
-        mShaderDefault.lightDirectional
-    )
-
-    private val managerLights = MGManagerLight(
-        mShaderDefault
-    )
-
-    private val mDrawerModeOpaque = MGDrawerModeOpaque(
-        mShaderSky,
+    private val mSceneTest = MGScene(
+        requesterUserContent,
         mShaderDefault,
-        mShaderWireframe,
-        meshSky,
-        mCameraFree,
-        mDrawerLightDirectional,
-        meshes,
-        mTriggers,
-        managerLights
+        mShaderSky,
+        mShaderNormals,
+        mShaderTexCoords,
+        mShaderWireframe
     )
-
-    private val mSwitcherDrawMode = MGSwitcherDrawMode(
-        meshSky,
-        meshes,
-        mDrawerModeOpaque
-    )
-
-    private val mLayerEditor = MGUILayerEditor(
-        clickLoadUserContent = MGClickGenerateLandscape(
-            mHandler,
-            mGeneratorLandscape,
-            requesterUserContent
-        ),
-        clickPlaceMesh = MGClickPlaceMesh(
-            mCallbackOnDeltaInteract,
-            meshes,
-            mDrawerSwitchBatch,
-            mShaderDefault,
-            mCallbackOnCameraMove
-        ),
-        clickSwitchDrawerMode = createDrawModeSwitcher()
-    ).apply {
-        setListenerTouchMove(
-            mCallbackOnCameraMove
-        )
-
-        setListenerTouchDelta(
-            mCallbackOnCameraMove
-        )
-
-        setListenerTouchScaleInteract(
-            object: MGIListenerScale {
-                override fun onScale(
-                    scale: Float
-                ) {
-                    mCallbackOnDeltaInteract.currentMeshInteract?.setScale(
-                        scale,
-                        scale,
-                        scale
-                    )
-                }
-            }
-        )
-
-        setListenerTouchDeltaInteract(
-            mCallbackOnDeltaInteract
-        )
-    }
-
-    private val mLight = MGLight(
-        MGVector(
-            1f, 1f, 0f
-        ),
-        MGVector(0f,0f,0f),
-        600f
-    ).apply {
-        managerLights.register(
-            this
-        )
-    }
-
-    private val mLight2 = MGLight(
-        MGVector(
-            1f, 0f, 1f
-        ),
-        MGVector(0f,0f,0f),
-        600f
-    ).apply {
-        managerLights.register(
-            this
-        )
-    }
     
     override fun onSurfaceCreated(
         gl: GL10?,
@@ -390,150 +178,8 @@ MGIListenerOnIntersectPosition {
             "shaders/texCoords/frag.glsl"
         )
 
-        val arrayVertexBox = MGArrayVertex().apply {
-            configure(
-                MGUtilsBuffer.createFloat(
-                    MGUtilsVertIndices.createCubeVertices(
-                        MGTriggerMethodBox.MIN,
-                        MGTriggerMethodBox.MAX
-                    )
-                ),
-                MGUtilsBuffer.createInt(
-                    MGUtilsVertIndices.createCubeIndices()
-                ),
-                stride = 3 * 4
-            )
-        }
-
-        val triggerAction = MGTriggerSimple(
-            mDrawerLightDirectional
-        )
-
-        MGObject3D.createFromAssets(
-            "objs/house.obj"
-        ).run {
-            mVerticesBatchObject.configure(
-                vertices,
-                indices
-            )
-
-            val modelMatrix = MGMatrixTransformationInvert(
-                MGMatrixScale()
-            )
-
-            val result = MGUtilsAlgo.findMinMaxPoints(
-                mVerticesBatchObject
-            )
-
-            modelMatrix.model.run {
-                setScale(
-                    result.second.x - result.first.x,
-                    result.second.y - result.first.y,
-                    result.second.z - result.first.z,
-                )
-                setPosition(
-                    (result.second.x + result.first.x) * 0.5f,
-                    (result.second.y + result.first.y) * 0.5f,
-                    (result.second.z + result.first.z) * 0.5f
-                )
-                invalidateScale()
-                invalidatePosition()
-                Log.d(TAG, "onSurfaceCreated: MODEL:")
-                modelMatrix.invert.calculateInvertModel()
-            }
-
-            mTriggers.add(
-                MGDrawerTriggerStateable(
-                    MGManagerTriggerState(
-                        MGTriggerMethodBox(
-                            modelMatrix.invert
-                        ),
-                        triggerAction
-                    ),
-                    arrayVertexBox,
-                    mShaderWireframe,
-                    modelMatrix.model
-                )
-            )
-        }
-
-        mTextureInteract.setupTexture(
-            "textures/rock.jpg"
-        )
-
-        mTextureLandscape.setupTexture(
-            "textures/terrain.png",
-            GL_REPEAT
-        )
-
-        mTextureSky.setupTexture(
-            "textures/sky/night.png"
-        )
-
-        MGObject3D.createFromAssets(
-            "objs/semi_sphere.obj"
-        ).run {
-            mVerticesSky.configure(
-                vertices,
-                indices
-            )
-        }
-
-        val landSize = 1024
-        mGeneratorLandscape.apply {
-            setResolution(
-                landSize,
-                landSize
-            )
-
-            val mapNormal = MGMapNormal.createFromAssets(
-                "maps/normal/terrain_normal.png"
-            )
-
-            forEachVertex(
-                MGVertexIteratorLandscapeNormal(
-                    mapNormal
-                )
-            )
-            mapNormal.destroy()
-
-            val mapHeight = MGMapDisplace.createFromAssets(
-                "maps/terrain_height.png"
-            )
-
-            forEachVertex(
-                MGVertexIteratorLandscapeDisplace(
-                    mapHeight,
-                    255 * 50,
-                    landSize,
-                    landSize
-                )
-            )
-            mapHeight.destroy()
-        }
-
-        val off = mGeneratorLandscape.actualWidth / -2f
-
-        modelMatrixLandscape.model.setPosition(
-            off,
-            -5500f,
-            off
-        )
-
-        modelMatrixLandscape.model.invalidatePosition()
-
-        val lx = 0.0f
-        val ly = 0.5f
-        val lz = -0.5f
-        val m = 2048f
-        modelMatrixCamera.setPosition(
-            0f, 0f, 0f
-        )
-        modelMatrixCamera.invalidatePosition()
-        mDrawerLightDirectional.setPosition(
-            2.8929129f,
-            10.986f,
-            -9.247298f
+        mSceneTest.onSurfaceCreated(
+            gl, config
         )
 
         glEnable(
@@ -557,19 +203,10 @@ MGIListenerOnIntersectPosition {
         Log.d(TAG, "onSurfaceChanged: ${Thread.currentThread().name}")
         mWidth = width
         mHeight = height
-
-        val fWidth = width.toFloat()
-        val fHeight = height.toFloat()
-
-        mCameraFree.setPerspective(
+        mSceneTest.onSurfaceChanged(
+            gl,
             width,
             height
-        )
-
-        mLayerEditor.layout(
-            0f, 0f,
-            fWidth,
-            fHeight
         )
     }
 
@@ -594,107 +231,22 @@ MGIListenerOnIntersectPosition {
             1.0f
         )
 
-        val t = System.currentTimeMillis() % 1000000L * 0.0007f
-
-        mLight.position.run {
-            x = cos(t) * 1250f
-            y = -1850f
-            z = -3250f
-        }
-
-        mLight2.position.run {
-            x = sin(t) * 1250f
-            y = -1850f
-            z = -3250f
-        }
-
         val error = glGetError()
         if (error != GL_NO_ERROR) {
             Log.d("TAG", "onDrawFrame: ERROR: ${error.toString(16)}")
             return
         }
 
-        // 1. Camera point triggering needs to check only on self position changes
-        // it doesn't need to check on each touch event
-        // 2. For other entities who can trigger, check it inside infinite loop
-        val model = mCameraFree.modelMatrix
-        mTriggers.forEach {
-            it.stateManager.trigger(
-                model.x - it.modelMatrix.x,
-                model.y - it.modelMatrix.y,
-                model.z - it.modelMatrix.z
-            )
-        }
-
-        mSwitcherDrawMode
-            .currentDrawerMode
-            .draw()
-
-        mHandler.run()
-    }
-
-    override fun onIntersectPosition(
-        p: MGVector
-    ) {
-        Log.d(TAG, "onIntersectPosition: ${p.x} ${p.y} ${p.z}")
-        mCallbackOnDeltaInteract.currentMeshInteract?.run {
-            setPosition(
-                p.x,
-                p.y,
-                p.z
-            )
-            invalidatePosition()
-        }
+        mSceneTest.onDrawFrame(
+            gl
+        )
     }
 
     fun onTouchEvent(
         event: MotionEvent
     ) {
-        mLayerEditor.onTouchEvent(
+        mSceneTest.onTouchEvent(
             event
         )
     }
-
-    private inline fun createDrawModeSwitcher() = MGClickSwitchDrawMode(
-        mHandler,
-        mSwitcherDrawMode,
-        MGMDrawMode(
-            mDrawerModeOpaque,
-            mShaderDefault,
-            mShaderSky,
-            mShaderDefault
-        ),
-        MGMDrawMode(
-            MGDrawerModeSingleShader(
-                mShaderWireframe,
-                meshSky,
-                mCameraFree,
-                meshes
-            ),
-            mShaderWireframe,
-            mShaderWireframe
-        ),
-        MGMDrawMode(
-            MGDrawerModeSingleShader(
-                mShaderNormals,
-                meshSky,
-                mCameraFree,
-                meshes
-            ),
-            mShaderNormals,
-            mShaderNormals,
-            mShaderNormals,
-            mShaderNormals
-        ),
-        MGMDrawMode(
-            MGDrawerModeSingleShader(
-                mShaderTexCoords,
-                meshSky,
-                mCameraFree,
-                meshes
-            ),
-            mShaderTexCoords,
-            mShaderTexCoords
-        )
-    )
 }
